+++ R/3_2_nutrient_data.R
@@ -79,7 +79,10 @@ nutrientDataServer <- function(cur_stn, has_focus) {
               inputId = ns("year"),
               label = NULL,
               choices = year_choices(cur_years()),
-              selected = last(cur_years())
+              selected = first_truthy(
+                intersect(isolate(input$year), cur_years()),
+                last(cur_years())
+              )
             )
           ),
           div(
@@ -109,7 +112,7 @@ nutrientDataServer <- function(cur_stn, has_focus) {
         tagList(
           div(
             class = "plot-caption",
-            "The dashed line on this plot indicates the total phosphorus state exceedance level of 0.075 mg/L (ppm). If more than one month of data was collected, the median and 80% confidence interval for the true total phosphorus level are displayed as a horizontal band."
+            "The dashed line on this plot indicates the total phosphorus state exceedance level of 0.075 mg/L (ppm). If more than one month of data was collected, the median and 80% confidence interval for the true total phosphorus level are displayed as a horizontal band. A zero value indicates the submitted sample was below the limit of detection."
           ),
           div(
             class = "plot-caption",